Extract audio commands into a separate module, simplify index.js
